refactorings

+++ app/src/main/java/ch/stephgit/memory/OverlayMessageFragment.kt
@@ -7,6 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.TextView
+import ch.stephgit.memory.ui.main.MainActivity
 
 class OverlayMessageFragment: Fragment() {
 
@@ -18,7 +19,7 @@ class OverlayMessageFragment: Fragment() {
 
         val view = inflater.inflate(R.layout.fragment_overlay_message, container, false)
         val textView = view.findViewById<TextView>(R.id.tv_overlay_message)
-        val message = "Ok! You had " + this.arguments?.getInt("flips")
+        val message = "Ok! You had " + this.arguments?.getLong("flips")
         textView.text =  message
 
         view.findViewById<Button>(R.id.btn_goto_newgame).setOnClickListener { startActivity(
