+++ app/src/main/java/ch/stephgit/memory/persistence/entity/Game.kt
@@ -5,6 +5,6 @@ import java.util.*
 data class Game (
     val username: String,
     val date: Date,
-    val flips: Int,
+    val flips: Long,
     var id: String? = null
 )
